Add bigInt support for large balances, fix balance display, add auto-mining for wallet creation

+++ pkg/blockchain/block.go
@@ -166,7 +166,7 @@ func SerializeTransactions(txs []*Transaction) []byte {
 		ID    string
 		From  string
 		To    string
-		Value float64
+		Value uint64
 		Data  []byte
 		Type  string
 	}
